refactor(console): review feedback — one component per file, drop restating comments
Extract WorkflowGraphFlow and WorkflowPanelFlow into their own files per the
one-JSX-component-per-file guideline; remove inline comments that restated
nearby prop JSDoc. The FlowNode vertical-label comment stays: it explains why
the label hangs beside the node (bottom edges would cross it), which the code
can't show.
