+++ packages/console/src/components/layout/ThreePaneLayout.tsx
@@ -78,7 +78,6 @@ export const ThreePaneLayout: React.FC<ThreePaneLayoutProps> = ({
           }}
         >
           <Box className={classes.listSurfaceCard} style={{ height: '100%', width: 'min(520px, 42vw)' }}>
-            {/* the center pane already draws the full workflow canvas */}
             <PanelContainer emptyMessage={emptyPanelMessage} workflowGraph={false} />
           </Box>
         </Box>
