+++ packages/console/src/components/panel/PanelFactory.tsx
@@ -13,8 +13,7 @@ import {
   useWorkflowContext,
   useWorkflowNode,
 } from '../../context/WorkflowContext'
-import { WorkflowGraphView } from '../project/WorkflowGraphView'
-import { PersonaTimeline } from '../flows/timeline/PersonaTimeline'
+import { WorkflowPanelFlow } from './WorkflowPanelFlow'
 import { useWorkflowInputSchema } from '../../hooks/useWorkflowInputSchema'
 import { FunctionTabbedPanel } from '../project/panels/FunctionDetailsForm'
 import {
@@ -185,26 +184,6 @@ const NewWorkflowRunForm: React.FC<{ workflowId: string }> = ({
         </Text>
       )}
     </Stack>
-  )
-}
-
-/* Vertical rendering of the workflow inside the (narrow) side panel: the
-   scenario timeline for scenarios, the top→down graph for everything else.
-   Suppressed (renderGraph=false) when the panel sits beside a full canvas
-   that already draws the same graph. */
-const WorkflowPanelFlow: React.FC = () => {
-  const { workflow } = useWorkflowContext()
-  const isScenario =
-    workflow?.source === 'scenario' || workflow?.scenario === true
-
-  return (
-    <Box h={480} style={{ minHeight: 0 }}>
-      {isScenario ? (
-        <PersonaTimeline workflow={workflow} />
-      ) : (
-        <WorkflowGraphView workflow={workflow} direction="DOWN" />
-      )}
-    </Box>
   )
 }
